// wash24-api/src/modules/auth/enums/auth.enums.ts
export enum SystemType {
  WASH24_WEB_POS = 'WASH24_WEB_POS',
  WASH24_MECHAN = 'WASH24',
  WASH24_ADMIN = 'WASH24_ADMIN',
  WASH24_APP = 'WASH24_APP',
  WASH24_DEVICE_CONTROLER = 'WASH24_DEVICE_CONTROLER',
}

export enum DeviceType {
  DESKTOP = 'DESKTOP',
  MOBILE = 'MOBILE',
  TABLET = 'TABLET',
  ESP32 = 'ESP32'
}